requirements.txt
streamlit
torch
torchvision
diffusers
transformers
accelerate
pillow